pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=2.2.13",
    "groq>=0.34.1",
    "mcp>=1.21.1",
    "python-dotenv>=1.0.0",
]
